Add a unit test for websocket wrapper and move integration tests into another file
